data-ingestion/producer/requirements.txt
beautifulsoup4==4.12.2
pymilvus==2.3.4 
langchain==0.1.0 
langchain-community==0.0.12 
langchain-core==0.1.10 
huggingface-hub==0.20.2
pulsar-client==3.4.0 
transformers[torch]==4.36.2
sentence-transformers==2.2.2  